Rename selection to ssm
