+++ doubleml/tests/test_ssm_exceptions.py
